🎊2.3.0
- SchedulerManager has been refactored from an abstract class to an interface
- The registerScheduler method in the SchedulerManager has been renamed to register
- The registerIndicator method in the IndicatorManager has been renamed to register
