+++ src/main/resources/messages/en.json
@@ -27,5 +27,5 @@
   "gui_next_page": "&aNext page",
   "gui_previous_page": "&aPrevious page",
   "generated_code": "&d&l----&b&lDARKWATERAPI&d&l----\n\n<COPY={CODE}::&b&lYour code:\n&7{CODE}>&a&nClick here to copy the code</COPY>\n\n",
-  "statistic": "&2Statistic\n&b-------------------------\n&a{STATISTIC}"
+  "statistic": "\n&2Statistic\n&b-------------------------\n&a{STATISTIC}"
 }
