+++ src/main/resources/messages/fr.json
@@ -27,5 +27,5 @@
   "gui_next_page": "&aPage suivante",
   "gui_previous_page": "&aPage précédente",
   "generated_code": "&d&l----&b&lDARKWATERAPI&d&l----\n\n<COPY={CODE}::&b&lVotre code :\n&7{CODE}>&a&nCliquez ici pour copier le code</COPY>\n\n " ,
-  "statistic": "&2Statistiques\n&b------------------------------\n&a{STATISTIC}"
+  "statistic": "\n&2Statistiques\n&b------------------------------\n&a{STATISTIC}"
 }
